exercise 2 corrected

+++ Assets/Scripts/EJ2.cs
@@ -4,16 +4,16 @@ using UnityEngine;
 //2.Solicitar el ingreso del precio de 3 productos y un monto de dinero disponible.Comunicar si la suma de los mismos supera o no el monto de dinero disponible.Luego indicar cuánto dinero sobra o falta.
 public class EJ2 : MonoBehaviour
 {
-	public int producto1;
-	public int producto2;
-	public int producto3;
-	public int monto;
+	public float producto1;
+	public float producto2;
+	public float producto3;
+	public float monto;
 	// Start is called before the first frame update
 	void Start()
     {
-		int suma;
-		int falta;
-		int resto;
+		float suma;
+		float falta;
+		float resto;
 		suma = producto1 + producto2 + producto3;
 		falta = suma - monto;
 		resto = suma - monto;
